+++ app/src/main/java/com/b1nd/alimo/presentation/feature/onboarding/parent/pw/second/ParentFindPWSecondFragment.kt
@@ -1,44 +1,79 @@
 package com.b1nd.alimo.presentation.feature.onboarding.parent.pw.second
 
+import android.os.CountDownTimer
+import android.view.View
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import com.b1nd.alimo.R
 import com.b1nd.alimo.databinding.FragmentParentFindPwSecondBinding
 import com.b1nd.alimo.presentation.base.BaseFragment
+import com.b1nd.alimo.presentation.feature.onboarding.parent.join.third.ParentJoinThirdViewModel
 import com.b1nd.alimo.presentation.feature.onboarding.parent.pw.second.ParentFindPWSecondViewModel.Companion.ON_CLICK_BACK
 import com.b1nd.alimo.presentation.feature.onboarding.parent.pw.second.ParentFindPWSecondViewModel.Companion.ON_CLICK_BACKGROUND
-import com.b1nd.alimo.presentation.feature.onboarding.parent.pw.second.ParentFindPWSecondViewModel.Companion.ON_CLICK_DONE
+import com.b1nd.alimo.presentation.feature.onboarding.parent.pw.second.ParentFindPWSecondViewModel.Companion.ON_CLICK_CERTIFICATION
+import com.b1nd.alimo.presentation.feature.onboarding.parent.pw.second.ParentFindPWSecondViewModel.Companion.ON_CLICK_NEXT
 import com.b1nd.alimo.presentation.utiles.hideKeyboard
 import com.b1nd.alimo.presentation.utiles.onSuccessEvent
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class ParentFindPWSecondFragment:
+class ParentFindPWSecondFragment :
     BaseFragment<FragmentParentFindPwSecondBinding, ParentFindPWSecondViewModel>(
-    R.layout.fragment_parent_find_pw_second
-) {
+        R.layout.fragment_parent_find_pw_second
+    ) {
     override val viewModel: ParentFindPWSecondViewModel by viewModels()
 
     override fun initView() {
+//        findNavController().popBackStack(R.id.parentLoginFirst, false)
+
         bindingViewEvent { event ->
             event.onSuccessEvent {
-                when(it){
+                when (it) {
                     ON_CLICK_BACK -> {
                         findNavController().popBackStack()
                     }
-                    ON_CLICK_DONE -> {
-                        findNavController().navigate(R.id.action_parentFindPwSecond_to_onBoardingThird)
-                    }
+
                     ON_CLICK_BACKGROUND -> {
                         mBinding.idEditTextLayout.clearFocus()
-                        mBinding.pwEditTextLayout.clearFocus()
+                        view?.hideKeyboard()
+                    }
+
+                    ON_CLICK_NEXT -> {
+                        findNavController().navigate(R.id.action_parentFindPWFirst_to_parentFindPWSecond)
+                    }
+
+                    ON_CLICK_CERTIFICATION -> {
+                        mBinding.check.visibility = View.VISIBLE
+                        mBinding.time.visibility = View.VISIBLE
+                        mBinding.certification.visibility = View.GONE
+
+
+                        object : CountDownTimer(300000, 1000) {
+
+                            override fun onTick(millisUntilFinished: Long) {
+                                val minutes = millisUntilFinished / 60000
+                                val seconds = (millisUntilFinished % 60000) / 1000
+                                val timeString = String.format("%d:%02d", minutes, seconds)
+                                mBinding.time.text = timeString
+                            }
+
+                            override fun onFinish() {
+                                mBinding.time.text = "0:00"
+                            }
+
+                        }.start()
+                    }
+
+                    ParentJoinThirdViewModel.ON_CLICK_CHECK -> {
+                        mBinding.nextBtnOff.visibility = View.INVISIBLE
+                        mBinding.nextBtnOn.visibility = View.VISIBLE
                         view?.hideKeyboard()
                     }
                 }
             }
         }
 
+        mBinding.checkLayout.bringToFront()
+
     }
-
-
 }
