+++ app/src/main/res/layout/fragment_parent_find_pw_first.xml
@@ -1,15 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
+
 <layout>
     <data>
         <variable
             name="vm"
-            type="com.b1nd.alimo.presentation.feature.onboarding.parent.pw.first.ParentFindPWFirstViewModel" />
+            type="com.b1nd.alimo.presentation.feature.onboarding.parent.pw.first." />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
-        android:onClick="@{() -> vm.onClickBackground()}"
         android:layout_height="match_parent">
 
         <LinearLayout
@@ -32,60 +32,24 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:padding="12dp"
-                android:text="@string/find_password"
-                android:textFontWeight="700"
+                android:text="@string/join"
                 android:textColor="@color/Main900"
+                android:textFontWeight="700"
                 android:textSize="16sp" />
         </LinearLayout>
 
         <androidx.appcompat.widget.AppCompatTextView
-            android:id="@+id/parent_join_thirst_title"
+            android:id="@+id/student_join_title"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginStart="24dp"
+            android:layout_marginStart="28dp"
             android:layout_marginTop="36dp"
+            android:text="@string/input_email"
+            android:textColor="@color/Main900"
             android:textFontWeight="700"
-            android:text="@string/parent_join_thirst_title"
             android:textSize="16sp"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/topBar" />
-
-        <LinearLayout
-            android:id="@+id/check_layout"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="10dp"
-            android:layout_marginEnd="30dp"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="@id/id_edit_text_layout">
-
-            <androidx.appcompat.widget.AppCompatTextView
-                android:id="@+id/time"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:padding="8dp"
-                android:text="5:00"
-                android:textColor="@color/Main500"
-                android:visibility="gone" />
-
-            <androidx.appcompat.widget.AppCompatImageButton
-                android:id="@+id/certification"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:background="@color/white"
-                android:onClick="@{() -> vm.onClickCertification()}"
-                android:src="@drawable/ic_certification" />
-
-            <androidx.appcompat.widget.AppCompatImageButton
-                android:id="@+id/check"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:background="@color/white"
-                android:onClick="@{() -> vm.onClickCheck()}"
-                android:src="@drawable/ic_check"
-                android:visibility="gone" />
-        </LinearLayout>
-
 
         <com.google.android.material.textfield.TextInputLayout
             android:id="@+id/id_edit_text_layout"
@@ -94,29 +58,31 @@
             android:layout_height="wrap_content"
             android:layout_marginStart="20dp"
             android:layout_marginTop="16dp"
-            app:hintEnabled="false"
             android:layout_marginEnd="20dp"
             android:textColorHint="@color/Gray500"
-            app:layout_constraintTop_toBottomOf="@id/parent_join_thirst_title">
+            app:endIconDrawable="@drawable/ic_clear_text"
+            app:endIconMode="custom"
+            app:hintEnabled="false"
+            app:layout_constraintTop_toBottomOf="@id/student_join_title">
 
             <com.google.android.material.textfield.TextInputEditText
                 android:id="@+id/id_edit_text"
-                android:maxLength="6"
                 android:layout_width="match_parent"
-                android:hint="@string/authentication_code"
-                android:textColorHint="@drawable/hint_color"
-
                 android:layout_height="52dp"
-                android:background="@drawable/edittext_border_background" />
+                android:background="@drawable/edittext_border_background"
+                android:hint="@string/parent_join_email"
+
+                android:textColorHint="@drawable/hint_color" />
 
         </com.google.android.material.textfield.TextInputLayout>
 
-
         <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/nextBtnOff"
+            android:id="@+id/loginBtnOff"
             style="?android:attr/borderlessButtonStyle"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
+            android:layout_marginStart="20dp"
+            android:layout_marginEnd="20dp"
             android:layout_marginBottom="30dp"
             android:background="@drawable/btn_shape_rounded_off"
             android:text="@string/next"
@@ -124,26 +90,24 @@
             android:visibility="visible"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            android:layout_marginStart="20dp"
-            android:layout_marginEnd="20dp"/>
+            app:layout_constraintStart_toStartOf="parent" />
 
         <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/nextBtnOn"
+            android:id="@+id/loginBtnOn"
             style="?android:attr/borderlessButtonStyle"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
+            android:layout_marginStart="20dp"
+            android:layout_marginEnd="20dp"
             android:layout_marginBottom="30dp"
             android:background="@drawable/btn_shape_rounded_on"
             android:onClick="@{() -> vm.onClickNext()}"
             android:text="@string/next"
             android:textColor="@color/Main900"
+            android:textFontWeight="700"
             android:visibility="invisible"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            android:layout_marginStart="20dp"
-            android:layout_marginEnd="20dp"/>
-
+            app:layout_constraintStart_toStartOf="parent" />
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
