+++ app/src/main/res/layout/fragment_parent_find_pw_second.xml
@@ -5,6 +5,7 @@
             name="vm"
             type="com.b1nd.alimo.presentation.feature.onboarding.parent.pw.second.ParentFindPWSecondViewModel" />
     </data>
+
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
@@ -31,99 +32,113 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:padding="12dp"
-                android:textFontWeight="700"
                 android:text="@string/find_password"
+                android:textFontWeight="700"
                 android:textColor="@color/Main900"
                 android:textSize="16sp" />
         </LinearLayout>
 
         <androidx.appcompat.widget.AppCompatTextView
-            android:id="@+id/parent_name"
+            android:id="@+id/parent_join_thirst_title"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="24dp"
             android:layout_marginTop="36dp"
             android:textFontWeight="700"
-            android:text="@string/parent_name"
-            android:textColor="@color/Main900"
+            android:text="@string/parent_join_thirst_title"
             android:textSize="16sp"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/topBar" />
 
-        <androidx.appcompat.widget.AppCompatTextView
-            android:id="@+id/student_join_title"
+        <LinearLayout
+            android:id="@+id/check_layout"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginStart="7dp"
-            android:textFontWeight="700"
+            android:layout_marginTop="10dp"
+            android:layout_marginEnd="30dp"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="@id/id_edit_text_layout">
 
-            android:layout_marginTop="36dp"
-            android:text="@string/parent_join_second_title"
-            android:textColor="@color/Main900"
-            android:textSize="16sp"
-            app:layout_constraintStart_toEndOf="@id/parent_name"
-            app:layout_constraintTop_toBottomOf="@id/topBar" />
+            <androidx.appcompat.widget.AppCompatTextView
+                android:id="@+id/time"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:padding="8dp"
+                android:text="5:00"
+                android:textColor="@color/Main500"
+                android:visibility="gone" />
+
+            <androidx.appcompat.widget.AppCompatImageButton
+                android:id="@+id/certification"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:background="@color/white"
+                android:onClick="@{() -> vm.onClickCertification()}"
+                android:src="@drawable/ic_certification" />
+
+            <androidx.appcompat.widget.AppCompatImageButton
+                android:id="@+id/check"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:background="@color/white"
+                android:onClick="@{() -> vm.onClickCheck()}"
+                android:src="@drawable/ic_check"
+                android:visibility="gone" />
+        </LinearLayout>
+
 
         <com.google.android.material.textfield.TextInputLayout
             android:id="@+id/id_edit_text_layout"
             style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox.Dense"
             android:layout_width="match_parent"
-            android:textFontWeight="700"
             android:layout_height="wrap_content"
             android:layout_marginStart="20dp"
             android:layout_marginTop="16dp"
-            android:layout_marginEnd="20dp"
             app:hintEnabled="false"
-
-            android:textColorHint="@color/Gray500"
-            app:endIconMode="password_toggle"
-            app:layout_constraintTop_toBottomOf="@id/student_join_title">
-
-            <com.google.android.material.textfield.TextInputEditText
-                android:layout_width="match_parent"
-                android:layout_height="52dp"
-                android:textColorHint="@drawable/hint_color"
-                android:hint="@string/new_password"
-                android:inputType="textPassword"
-                android:background="@drawable/edittext_border_background" />
-
-        </com.google.android.material.textfield.TextInputLayout>
-
-        <com.google.android.material.textfield.TextInputLayout
-            android:id="@+id/pw_edit_text_layout"
-            style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox.Dense"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="20dp"
-            android:layout_marginTop="8dp"
-            app:hintEnabled="false"
-
             android:layout_marginEnd="20dp"
             android:textColorHint="@color/Gray500"
-            app:endIconMode="password_toggle"
-            app:layout_constraintTop_toBottomOf="@id/id_edit_text_layout">
+            app:layout_constraintTop_toBottomOf="@id/parent_join_thirst_title">
 
             <com.google.android.material.textfield.TextInputEditText
+                android:id="@+id/id_edit_text"
+                android:maxLength="6"
                 android:layout_width="match_parent"
-                android:layout_height="52dp"
+                android:hint="@string/authentication_code"
                 android:textColorHint="@drawable/hint_color"
-                android:hint="@string/re_new_password"
-                android:inputType="textPassword"
+
+                android:layout_height="52dp"
                 android:background="@drawable/edittext_border_background" />
 
         </com.google.android.material.textfield.TextInputLayout>
 
 
         <androidx.appcompat.widget.AppCompatButton
+            android:id="@+id/nextBtnOff"
+            style="?android:attr/borderlessButtonStyle"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="30dp"
+            android:background="@drawable/btn_shape_rounded_off"
+            android:text="@string/next"
+            android:textColor="@color/Gray500"
+            android:visibility="visible"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginStart="20dp"
+            android:layout_marginEnd="20dp"/>
+
+        <androidx.appcompat.widget.AppCompatButton
+            android:id="@+id/nextBtnOn"
             style="?android:attr/borderlessButtonStyle"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_marginBottom="30dp"
             android:background="@drawable/btn_shape_rounded_on"
-            android:text="@string/complete"
-            android:onClick="@{() -> vm.onClickDone()}"
+            android:onClick="@{() -> vm.onClickNext()}"
+            android:text="@string/next"
             android:textColor="@color/Main900"
-            android:textFontWeight="700"
+            android:visibility="invisible"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
